requirements.txt
langchain-upstage==0.1.7
langgraph==0.2.4
streamlit==1.37.1
python-dotenv==1.0.1
langchain==0.2.14